hotel-review-sentiment-analyzer/requirements.txt
ibm-watson-machine-learning
python-dotenv
pandas
streamlit
schedule